PDBUtils.py with list with ligands, PrePrepare is working

+++ Code/DOCK_Prepare.py
@@ -1,6 +1,5 @@
 #Written by Daniel Zaidman
 #Code review by 
-
 import shutil
 import subprocess
 import os
@@ -42,5 +41,7 @@ class DOCK_Prepare:
         os.rename(INDOCK + '2', INDOCK)
 #Inner functions
     def create_fixed_names(self):
-        shutil.copyfile(self.rec, self.fixed_rec)
-        shutil.copyfile(self.lig, self.fixed_lig)
+        if(not os.path.exists(self.fixed_rec)):
+            shutil.copyfile(self.rec, self.fixed_rec)
+        if(not os.path.exists(self.fixed_lig)):
+            shutil.copyfile(self.lig, self.fixed_lig)
